user specific content with jwt claims

+++ Tweetbook/Controllers/V1/PostsController.cs
@@ -7,6 +7,7 @@ using Tweetbook.Contracts.V1;
 using Tweetbook.Contracts.V1.Requests;
 using Tweetbook.Contracts.V1.Responses;
 using Tweetbook.Domain;
+using Tweetbook.Extensions;
 using Tweetbook.Services;
 
 namespace Tweetbook.Controllers.V1
@@ -41,7 +42,8 @@ namespace Tweetbook.Controllers.V1
         [HttpPost(ApiRoutes.Posts.Create)]
         public async Task<IActionResult> Create(CreatePostRequest postRequest)
         {
-            var post = new Post { Name = postRequest.Name };
+            var userId = HttpContext.GetUserId();
+            var post = new Post { Name = postRequest.Name, UserId = userId};
             
             await _postService.CreatePostAsync(post);
             var locationUrl = HttpContext.Request.Scheme + "://" + HttpContext.Request.Host
@@ -54,7 +56,14 @@ namespace Tweetbook.Controllers.V1
         [HttpPut(ApiRoutes.Posts.Update)]
         public async Task<IActionResult> Update([FromRoute] Guid postId, [FromBody] UpdatePostRequest request)
         {
-            var post = new Post { Id = postId, Name = request.Name };
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+            if (!userOwnsPost)
+            {
+                return BadRequest(new {error = "You do not own this post."});
+            }
+
+            var post = await _postService.GetPostByIdAsync(postId);
+            post.Name = request.Name;
             var updated = await _postService.UpdatePostAsync(post);
             if (updated)
                 return Ok();
@@ -65,6 +74,11 @@ namespace Tweetbook.Controllers.V1
         [HttpDelete(ApiRoutes.Posts.Delete)]
         public async Task<IActionResult> Delete([FromRoute] Guid postId)
         {
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+            if (!userOwnsPost)
+            {
+                return BadRequest(new {error = "You do not own this post."});
+            }
             bool deleted = await _postService.DeleteAsync(postId);
             if (deleted) return NoContent();
             return NotFound();
